Separate model catalog from registry and reorganize provider schemas.
Ship catalog YAML under catalog/ with MODEL_CATALOG_PATH, split catalog vs configured-provider Zod ownership under schemas/, and tighten NameSchema to SF-aligned slug rules.
