+++ docker-compose.yml
@@ -58,7 +58,8 @@ services:
       - path: packages/server/.env
         required: true
     # The YAML registry (models.yaml, mcp.yaml, skills.yaml) is not baked into
-    # the image. Copy registry-example to registry before starting the service.
+    # the image (the shipped model catalog under catalog/ is). Copy
+    # registry-example to registry before starting the service.
     volumes:
       - ./packages/server/registry:/app/packages/server/registry:ro
     depends_on:
