+++ packages/server/.env.example
@@ -14,6 +14,10 @@ PORT=8790
 # in models.yaml; only its API key lives here.
 # REGISTRY_DIR=registry
 
+# Shipped model catalog YAML (discovery presets for GET /model-providers/catalog).
+# Separate from REGISTRY_DIR. Optional, defaults to ./catalog/model-catalog.yaml.
+# MODEL_CATALOG_PATH=catalog/model-catalog.yaml
+
 # Frontend build the server serves alongside the API. Optional, defaults to
 # ../frontend/dist relative to the working directory. With no build there the
 # server logs a warning and serves the API only.
