+++ packages/server/src/catalog/ModelCatalog.ts
@@ -3,9 +3,9 @@
  * and model presets the settings UI copies into PUT /model-providers bodies.
  * Never consulted on writes and never read by the runtime.
  */
-import { CONFIG_FILES } from '../config';
-import { loadYamlFile } from '../store/loadYaml';
-import { ModelCatalogFileSchema, type CatalogProvider } from './schemas';
+import configuration from '../config';
+import { ModelCatalogFileSchema, type CatalogProvider } from '../schemas/modelCatalog';
+import { loadYamlAtPath } from './loadYaml';
 
 export class ModelCatalog {
   private readonly providers: CatalogProvider[];
@@ -14,9 +14,9 @@ export class ModelCatalog {
     this.providers = providers;
   }
 
-  /** Loads and validates model-catalog.yaml from `REGISTRY_DIR`. Throws on any error. */
+  /** Loads and validates the catalog at `MODEL_CATALOG_PATH`. Throws on any error. */
   static load(): ModelCatalog {
-    const file = loadYamlFile(CONFIG_FILES.modelCatalog, ModelCatalogFileSchema);
+    const file = loadYamlAtPath(configuration.MODEL_CATALOG_PATH, ModelCatalogFileSchema);
     return new ModelCatalog(file.providers);
   }
