+++ packages/server/src/db/sqlite/model-provider-store/SqliteModelProviderStore.ts
@@ -1,5 +1,6 @@
 import type { ExpressionBuilder, Kysely } from 'kysely';
-import type { ProviderManifest, ResourceName } from '../../../catalog/schemas';
+import type { ResourceName } from '../../../schemas/common';
+import type { ProviderManifest } from '../../../schemas/modelProvider';
 import {
   flattenProviderModels,
   type IModelProviderStore,
