+++ packages/server/src/db/sqlite/types.ts
@@ -17,7 +17,7 @@ import type {
 } from '@truefoundry/utils/core';
 import type { CurrentContextUsage } from '@truefoundry/utils/core/runtime/contextUsage';
 import type { ColumnType, Generated, JSONColumnType } from 'kysely';
-import type { ProviderManifest } from '../../catalog/schemas';
+import type { ProviderManifest } from '../../schemas/modelProvider';
 import type { McpServerManifest } from '../../store/schemas';
 
 /**
